+++ src/affiliate-manager/view/affManager.js
@@ -22,7 +22,6 @@ export class AffManager {
 
   stateChanged(newState) {
     this.showError = !!newState.error.errorMessage;
-    console.log(this.showError)
   }
 
   unbind() {
